# docs/swagger.yaml
basePath: /api
definitions:
  handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  models.CreateNodeDependencyRequest:
    properties:
      cascade_delete:
        type: boolean
      cascade_update:
        type: boolean
      dependency_node_id:
        type: integer
      dependency_type:
        enum:
        - hard
        - soft
        - reference
        type: string
      metadata:
        $ref: '#/definitions/models.DependencyMetadata'
    required:
    - dependency_node_id
    - dependency_type
    type: object
...
  models.DependencyMetadata:
    properties:
      description:
        type: string
      relationship:
        type: string
    type: object
...
  models.EventChanges:
    properties:
      after:
        additionalProperties: true
        type: object
      before:
        additionalProperties: true
        type: object
    type: object
  models.EventData:
    properties:
      changes:
        $ref: '#/definitions/models.EventChanges'
      event_id:
        type: string
      event_type:
        type: string
      metadata:
        additionalProperties: true
        type: object
      node_id:
        type: integer
      timestamp:
        type: string
    type: object
...
  models.NodeDependency:
    properties:
      cascade_delete:
        type: boolean
      cascade_update:
        type: boolean
      created_at:
        type: string
      dependency_node_id:
        type: integer
      dependency_type:
        type: string
      dependent_node_id:
        type: integer
      id:
        type: integer
      metadata:
        $ref: '#/definitions/models.DependencyMetadata'
    type: object
  models.NodeEvent:
    properties:
      event_data:
        $ref: '#/definitions/models.EventData'
      event_type:
        type: string
      id:
        type: integer
      node_id:
        type: integer
      occurred_at:
        type: string
      processed_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A URL management system with attribute support and MCP integration.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: URL Database API
  version: "1.0"
paths:
  /api/dependencies/{id}:
    delete:
      description: Delete a dependency relationship
      parameters:
      - description: Dependency ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete dependency
      tags:
      - dependencies
    get:
      description: Get a dependency by ID
      parameters:
      - description: Dependency ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NodeDependency'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get dependency
      tags:
      - dependencies
  /api/events:
    get:
      description: Get events by type within a date range
      parameters:
      - description: Event type
        in: query
        name: type
        required: true
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start
        required: true
        type: string
      - description: End date (RFC3339)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NodeEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get events by type
      tags:
      - events
  /api/events/{eventId}/process:
    post:
      description: Mark an event as processed
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Process event
      tags:
      - events
  /api/events/cleanup:
    post:
      description: Delete processed events older than specified days
      parameters:
      - description: Number of days to retain events
        in: query
        minimum: 1
        name: days
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Cleanup old events
      tags:
      - events
  /api/events/pending:
    get:
      description: Get unprocessed events for processing
      parameters:
      - default: 100
        description: Maximum number of events to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NodeEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get pending events
      tags:
      - events
  /api/events/stats:
    get:
      description: Get statistics about events in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get event statistics
      tags:
      - events
  /api/nodes/{nodeId}/dependencies:
    get:
      description: Get all dependencies where the node is dependent
      parameters:
      - description: Node ID
        in: path
        name: nodeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NodeDependency'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get node dependencies
      tags:
      - dependencies
    post:
      consumes:
      - application/json
      description: Create a new dependency relationship between nodes
      parameters:
      - description: Dependent Node ID
        in: path
        name: nodeId
        required: true
        type: integer
      - description: Dependency details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateNodeDependencyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.NodeDependency'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Circular dependency detected
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create dependency
      tags:
      - dependencies
  /api/nodes/{nodeId}/dependents:
    get:
      description: Get all nodes that depend on this node
      parameters:
      - description: Node ID
        in: path
        name: nodeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NodeDependency'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get node dependents
      tags:
      - dependencies
  /api/nodes/{nodeId}/events:
    get:
      description: Get events for a specific node
      parameters:
      - description: Node ID
        in: path
        name: nodeId
        required: true
        type: integer
      - default: 50
        description: Maximum number of events to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.NodeEvent'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get node events
      tags:
      - events
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"